Fixed comm noise action
Unbounded action from neural network gaussian leads to negative values that make (1+noise)<0, and thus making the sum of comm vals unusual. To fix this, we abs() the action so that only lying above the singleton value is possible.

Also, used self. variables where possible and added a frac error to flag if the error ever pops up again.