discretised the comm action space
The communication action space was continuous, but this was causing some convergence issues and divide by zero errors. It also was difficult to bound the actions and explore that region more thoroughly. This meant that convergence was much worse.

Added a controller to control how finely the range is divided.

Also added some test results to control.